estimate combine items, and added a separate dialog

+++ qt4gui/customwidgets/estimateWidget.py
@@ -12,16 +12,18 @@ import re
 from PyQt4 import QtGui, QtCore
 from openmolar.qt4gui.customwidgets import Ui_estimateItemWidget,\
 Ui_estHeaderWidget,Ui_estFooterWidget,chainLabel
-
+from openmolar.qt4gui.dialogs import Ui_estSplitItemsDialog
 
 class estItemWidget(Ui_estimateItemWidget.Ui_Form):
     def __init__(self,parent=None):
         self.parent=parent
         self.setupUi(self.parent)
         self.addchain()
-        self.signals()
         self.validators()
         self.feesLinked=True
+        self.items=[]
+        self.itemCode=""
+        self.signals()
         
     def addchain(self):
         self.chain=chainLabel.chainLabel(self.chain_frame)
@@ -37,31 +39,66 @@ class estItemWidget(Ui_estimateItemWidget.Ui_Form):
     def setChain(self,cset):
         if cset!="P":
             self.chain.mousePressEvent(None)
+
     def setItem(self,item):
-        self.item=item
+        self.items.append(item)
         self.code_label.setToolTip(self.toolTip())
-    
+        self.loadValues()
+        
+    def separateIcon(self):
+        icon = QtGui.QIcon()
+        icon.addPixmap(QtGui.QPixmap(":icons/separate.png"), 
+        QtGui.QIcon.Normal, QtGui.QIcon.Off)
+
+        self.delete_pushButton.setIcon(icon)
+        
+    def addItem(self,item):
+        self.items.append(item)
+        self.code_label.setToolTip(self.toolTip())
+        self.multiValues()
+        self.separateIcon()
+        self.code_label.setText("multi")
+        
     def toolTip(self):
-        return '''<center>Type - '%s'<br />ItemCode - '%s'<br />Feescale - %s
-        <br />CSEtype - %s<br />Dent - %s</center>DBindex - %s'''%(
-        self.item.type,
-        self.item.itemcode,
-        self.item.feescale,
-        self.item.csetype,
-        self.item.dent,
-        self.item.ix)
-                
+        retarg='<center>'
+        for item in self.items:
+            retarg+='''Type - '%s'<br />ItemCode - '%s'<br />Feescale - %s
+            <br />CSEtype - %s<br />Dent - %s<br />DBindex - %s<hr />'''%(
+            item.type,
+            item.itemcode,
+            item.feescale,
+            item.csetype,
+            item.dent,
+            item.ix)
+        return retarg+"</center>"
+    
     def loadValues(self):
-        if self.item.number!=None:
-            self.setNumber(self.item.number)
-        self.setFee(self.item.fee)
-        self.setPtFee(self.item.ptfee)
-        self.setDescription(self.item.description)
-        self.setType(self.item.type)
-        self.setItemCode(self.item.itemcode)
-        self.setCompleted(self.item.completed)
-        self.setCset(self.item.csetype)
-        self.setChain(self.item.csetype)
+        item =self.items[0]
+        
+        if item.number!=None:
+            self.setNumber(item.number)
+        self.setFee(item.fee)
+        self.setPtFee(item.ptfee)
+        self.setDescription(item.description)
+        self.setType(item.type)
+        self.setItemCode(item.itemcode)
+        self.connectCB(False)
+        self.setCompleted(item.completed)
+        self.connectCB()
+        self.setCset(item.csetype)
+        self.setChain(item.csetype)
+
+    def multiValues(self):
+        print "multiValues"
+        fee,ptfee,number=0,0,0
+        for item in self.items:
+            if item.number:
+                number+=item.number
+            fee+=item.fee
+            ptfee+=item.ptfee
+        self.setNumber(number)
+        self.setFee(fee)
+        self.setPtFee(ptfee)
         
     def validators(self):
         self.fee_lineEdit.setValidator(QtGui.\
@@ -72,13 +109,24 @@ class estItemWidget(Ui_estimateItemWidget.Ui_Form):
         #self.fee_lineEdit.setInputMask('000.00')
         #self.ptFee_lineEdit.setInputMask("000.00")
 
+    def connectCB(self,connect=True):
+        '''
+        connects the signals from the completed checkbox
+        to stop programmatic firing
+        '''
+        if connect:
+            QtCore.QObject.connect(self.completed_checkBox,
+            QtCore.SIGNAL("stateChanged(int)"),self.completeItem)
+        else:
+            QtCore.QObject.disconnect(self.completed_checkBox,
+            QtCore.SIGNAL("stateChanged(int)"),self.completeItem)
+            
     def signals(self):
         QtCore.QObject.connect(self.delete_pushButton,
         QtCore.SIGNAL("clicked()"),self.deleteItem)
 
-        QtCore.QObject.connect(self.completed_checkBox,
-        QtCore.SIGNAL("stateChanged(int)"),self.completeItem)
-
+        self.connectCB()
+        
         self.cset_lineEdit.connect(self.cset_lineEdit,QtCore.SIGNAL(
         "textEdited (const QString&)"),self.update_cset)
 
@@ -113,7 +161,8 @@ class estItemWidget(Ui_estimateItemWidget.Ui_Form):
                 self.update_ptFee(arg,False)
         except ValueError:
             newVal=0
-        self.item.fee=newVal
+        for item in self.items:
+            item.fee=newVal/len(self.items)
         if userPerforming:
             self.userInput()
     
@@ -125,7 +174,8 @@ class estItemWidget(Ui_estimateItemWidget.Ui_Form):
                 self.update_Fee(arg,False)
         except ValueError:
             newVal=0
-        self.item.ptfee=newVal
+        for item in self.items:
+            item.ptfee=newVal/len(self.items)
         if userPerforming:
             self.userInput()
     def setNumber(self,arg):
@@ -136,9 +186,15 @@ class estItemWidget(Ui_estimateItemWidget.Ui_Form):
 
     def setType(self,arg):
         if arg in (None,""):
-            arg="-"
-        self.code_label.setText(arg.split(" ")[0])
+            self.code_label.setText("?")
+        elif arg[:2] in ("ur","ul","ll","lr"):
+            arg="(%s)"%arg.split(" ")[0].upper()
+            self.code_label.setText(arg)
+        else:
+            self.code_label.setText(arg.split(" ")[0])
+    
     def setItemCode(self,arg):
+        self.itemcode=arg
         if arg in (None,""):
             arg="-"
         self.itemCode_label.setText(arg)
@@ -158,27 +214,72 @@ class estItemWidget(Ui_estimateItemWidget.Ui_Form):
         '''
         a slot for the delete button press
         '''
-        self.parent.emit(QtCore.SIGNAL("deleteMe"),(self))
+        if len(self.items)>1:
+            self.splitMultiItemDialog()
+        else:
+            self.parent.emit(QtCore.SIGNAL("deleteMe"),(self))
 
     def completeItem(self,arg):
         '''
         a slot for the checkbox state change
         should only happen when this is altered by user (not programatically)
         '''
+        print "completing Item"
         result=(arg==2)
         #self.delete_pushButton.setEnabled(not result)
         self.fee_lineEdit.setEnabled(not result)
         self.ptFee_lineEdit.setEnabled(not result)
-
-        self.item.completed=result
+        
+        self.items[0].completed=result
         self.userInput()
-        self.parent.emit(QtCore.SIGNAL("completedItem"), (self.item))
+        self.parent.emit(QtCore.SIGNAL("completedItem"), (self.items[0]))
 
     def userInput(self):
         self.parent.emit(QtCore.SIGNAL("user_interaction"))
 
+    def splitMultiItemDialog(self):
+        Dialog=QtGui.QDialog(self.parent)
+        dl=Ui_estSplitItemsDialog.Ui_Dialog()
+        dl.setupUi(Dialog)
+        ew=estWidget()
+        ew.setEstimate(self.items,True)
+        dl.scrollArea.setWidget(ew)
+        
+        #-- this miniDialog emits signals that go uncaught 
+        ew.connect(ew, QtCore.SIGNAL("applyFeeNow"), self.passOnApplyFeeNowSignal)
+        ew.connect(ew, QtCore.SIGNAL("completedItem"), self.passOnCompletedSignal)
+        ew.connect(ew, QtCore.SIGNAL("unCompletedItem"), self.passOnUncompletedSignal)
+        ew.connect(ew, QtCore.SIGNAL("deleteItem"), self.passOnDeleteItemSignal)
 
-
+        Dialog.exec_()
+        ##refresh NOW!!!
+        
+    def passOnCompletedSignal(self,arg):
+        '''
+        the child dialog has emitted a signal... pass it on
+        '''
+        self.parent.parent().emit(QtCore.SIGNAL("completedItem"),arg)
+    def passOnUncompletedSignal(self,arg):
+        '''
+        the child dialog has emitted a signal... pass it on
+        '''
+        self.parent.parent().emit(QtCore.SIGNAL("unCompletedItem"),arg)
+    def passOnApplyFeeNowSignal(self,arg):
+        '''
+        the child dialog has emitted a signal... pass it on
+        '''
+        self.parent.parent().emit(QtCore.SIGNAL("applyFeeNow"),arg)
+    def passOnDeleteItemSignal(self,arg):
+        '''
+        the child dialog has emitted a signal... pass it on
+        '''
+        self.parent.parent().emit(QtCore.SIGNAL("deleteItem"),arg)
+        
+    
+    
+    def signalPass(self,arg=None):
+        print "signal caught",arg
+            
 class estWidget(QtGui.QFrame):
     def __init__(self,parent=None):
         super(estWidget,self).__init__(parent)
@@ -221,24 +322,35 @@ class estWidget(QtGui.QFrame):
         self.estFooter.fee_lineEdit.setText("%.02f"%(self.total/100))
         self.estFooter.ptfee_lineEdit.setText("%.02f"%(self.ptTotal/100))
 
-    def setEstimate(self,ests):
+    def findExistingItemWidget(self,item):
+        for widg in self.estItemWidgets:
+            if widg.itemcode==item.itemcode:
+            #and widg.description == item.description:
+                widg.addItem(item)
+                return True
+                
+    def setEstimate(self,ests,SPLIT_ALL=False):
         self.ests=ests
         self.clear()
         for item in self.ests:
-            iw=QtGui.QWidget()
-            i=estItemWidget(iw)
-            i.setItem(item)
-            i.loadValues()
-            iw.connect(iw,QtCore.SIGNAL("user_interaction"),
-            self.updateTotals)
+            if not SPLIT_ALL and self.findExistingItemWidget(item):
+                #-- try and match with existing items
+                print "added est to an existing widget item"
+            else:
+                #--creats a widget
+                iw=QtGui.QWidget(self)
+                i=estItemWidget(iw)
+                i.setItem(item)
+                iw.connect(iw,QtCore.SIGNAL("user_interaction"),
+                self.updateTotals)
 
-            iw.connect(iw,QtCore.SIGNAL("deleteMe"),self.deleteItemWidget)
+                iw.connect(iw,QtCore.SIGNAL("deleteMe"),self.deleteItemWidget)
 
-            iw.connect(iw,QtCore.SIGNAL("completedItem"),
-            self.itemCompletionState)
-            self.estItemWidgets.append(i)
-            self.estimate_layout.insertWidget(1,iw)
+                iw.connect(iw,QtCore.SIGNAL("completedItem"),
+                self.itemCompletionState)
 
+                self.estItemWidgets.append(i)
+                self.estimate_layout.insertWidget(1,iw)
         self.setMinimumSize(self.minimumSizeHint())
         self.updateTotals()
 
@@ -277,7 +389,7 @@ class estWidget(QtGui.QFrame):
 
             arg.parent.setParent(None)
             for est in self.ests:
-                if est.ix==arg.item.ix and est.type==arg.item.type:
+                if est.ix==arg.items[0].ix and est.type==arg.items[0].type:
                     self.ests.remove(est)
                     self.emit(QtCore.SIGNAL("deleteItem"),est.type)
